pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
django = "5.0.3"
sqlparse = "^0.5.0"
amqp = "5.1.1"
annotated-types = "0.6.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
billiard = "4.1.0"
black = "24.3.0"
celery = "5.3.4"
certifi = "2023.7.22"
cffi = "1.15.1"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cryptography = "^42.0.5"
django-cors-headers = "4.3.1"
django-ninja = "1.1.0"
exceptiongroup = "1.1.3"
h11 = "0.14.0"
idna = "^3.7"
kombu = "5.3.2"
mypy-extensions = "1.0.0"
outcome = "1.2.0"
packaging = "24.0"
pathspec = "0.12.1"
platformdirs = "4.2.0"
prompt-toolkit = "3.0.39"
psycopg = "3.1.10"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
pyjwt = "2.8.0"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
redis = "5.0.1"
selenium = "4.13.0"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
trio = "0.22.2"
trio-websocket = "0.11.1"
typing-extensions = "4.10.0"
tzdata = "2023.3"
urllib3 = "^2.2.1"
vine = "5.0.0"
wcwidth = "0.2.7"
wsproto = "1.2.0"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
